--- bettermemory/src/main/java/no/bettermemory/interfaces/MicrocontrollerDatabaseBridge/ObjectQue.java
@@ -1,7 +1,0 @@
-package no.bettermemory.interfaces.MicrocontrollerDatabaseBridge;
-
-public interface ObjectQue<T, U> {
-    int addToQue(T listOrMapObject);
-    U getFirstFromQueAsDTO();
-    boolean containsElements();
-}
