--- bettermemory/src/main/java/no/bettermemory/interfaces/MicrocontrollerDatabaseBridge/TimeBasedObjectRetriever.java
@@ -1,16 +1,0 @@
-package no.bettermemory.interfaces.MicrocontrollerDatabaseBridge;
-
-/**
- * This interface is meant to be used in order to return a 
- * an activity based on some set of time parameters.
- */
-public interface TimeBasedObjectRetriever<T> {
-
-    /**
-     * Should return the activity in a chosen data format.
-     * @return T
-     * 
-     */
-    T getObject(); 
-    
-}
