+++ bettermemory/src/main/java/no/bettermemory/interfaces/storageHandlers/storageGetters/GetActivity.java
@@ -12,8 +12,8 @@ public interface GetActivity {
         String patientId, int year, int weekNumber, String dayName, int hour, int minutes
     ) throws Exception;
 
-    HashMap<ObjectId, Activity> getActivitiesAtInterval(
-        String patientId, int year, int weekNumber, String dayName, int currentHour, int currentMinutes, int interval
+    HashMap<ObjectId, Activity> getActivitiesAtHour(
+        String patientId, int year, int weekNumber, String dayName, int hour
     ) throws Exception;
     
     HashMap<ObjectId, Activity> getActivitiesFromObjectId(List<ObjectId> activityIds) throws Exception;
