--- bettermemory/src/main/java/no/bettermemory/models/MicrocontrollerDatabaseBridge/ActivityHandlers/ArrayDTOHandler.java
@@ -1,7 +1,0 @@
-package no.bettermemory.models.MicrocontrollerDatabaseBridge.ActivityHandlers;
-
-public class ArrayDTOHandler<T> extends ArrayHandler<T> {
-    public ArrayDTOHandler(T[] array) {
-        super(array);
-    }
-}
