+++ bettermemory/src/main/java/no/bettermemory/models/MicrocontrollerDatabaseBridge/PresentWeekInteger.java
@@ -1,8 +1,10 @@
-package no.bettermemory.models.MicrocontrollerDatabaseBridge.PresentTimeClasses;
+package no.bettermemory.models.MicrocontrollerDatabaseBridge;
 
-import no.bettermemory.interfaces.Models.MinutesProvider;
+import no.bettermemory.interfaces.Models.WeekProvider;
 import java.time.LocalDateTime;
 import java.time.Clock;
+import java.util.Locale;
+import java.time.temporal.WeekFields;
 
 /**
  * This class is meant to provide the current minutes. However, it is possible to
@@ -14,7 +16,7 @@ import java.time.Clock;
  * @author 
  * Hermann Mjelde Hamnnes
  */
-public class PresentMinutesInteger implements MinutesProvider<Integer> {
+public class PresentWeekInteger implements WeekProvider<Integer> {
 
     private final Clock clock;
 
@@ -22,7 +24,7 @@ public class PresentMinutesInteger implements MinutesProvider<Integer> {
      * If no Clock object was passed to the constructor, the clock will
      * by default be set to the systemDefaultZone.
      */
-    public PresentMinutesInteger() {
+    public PresentWeekInteger() {
         this.clock = Clock.systemDefaultZone();
     }
 
@@ -31,7 +33,7 @@ public class PresentMinutesInteger implements MinutesProvider<Integer> {
      * a different Clock than the systemDefaultZone.
      * @param Clock
      */
-    public PresentMinutesInteger(Clock clock) {
+    public PresentWeekInteger(Clock clock) {
         this.clock = clock;
     }
 
@@ -39,10 +41,15 @@ public class PresentMinutesInteger implements MinutesProvider<Integer> {
     @Override
     /**
      * This class will provide an integer which represent the current or a predefined 
-     * minutes.
+     * weekNumber.
      */
-    public Integer getMinutes() {
-        return  LocalDateTime.now(clock).getMinute();
+    public Integer getWeek() {
+        return  LocalDateTime.now(clock) //sets the clock for this the returned now value.
+                             .get(WeekFields.of(Locale.getDefault()) 
+                             .weekOfWeekBasedYear()); /*Sets the returned week number to 
+                                                  be based on the given year in the returned
+                                                  now value.
+                                                */
         
     }
