+++ bettermemory/src/test/java/no/bettermemory/models/MicrocontrollerDatabaseBridge/PresentDayStringTest.java
@@ -6,7 +6,6 @@ import org.junit.jupiter.api.extension.ExtendWith;
 
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import no.bettermemory.models.MicrocontrollerDatabaseBridge.PresentTimeClasses.PresentDayString;
 
 import java.time.Clock;
 import java.time.Instant;
