+++ bettermemory/src/test/java/no/bettermemory/models/MicrocontrollerDatabaseBridge/PresentMinutesIntegerTest.java
@@ -6,7 +6,6 @@ import org.junit.jupiter.api.extension.ExtendWith;
 
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import no.bettermemory.models.MicrocontrollerDatabaseBridge.PresentTimeClasses.PresentMinutesInteger;
 
 import java.time.Clock;
 import java.time.Instant;
